requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.10.0
PyMySQL==1.1.0
Werkzeug==2.3.7
WTForms==3.0.1
email_validator==2.1.1
Pillow==11.2.1
